Refactored most charts into separate files
