+++ src/parser/mage/arcane/modules/ManaChart/TabComponent/ManaChart.scss
