+++ src/parser/mage/arcane/modules/ManaChart/TabComponent/ManaLevelGraph.js
@@ -1,137 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {
-  FlexibleWidthXYPlot as XYPlot,
-  DiscreteColorLegend,
-  XAxis,
-  YAxis,
-  VerticalGridLines,
-  HorizontalGridLines,
-  AreaSeries,
-  LineSeries,
-} from 'react-vis';
 
 import fetchWcl from 'common/fetchWclApi';
 
 import ManaStyles from 'interface/others/ManaStyles.js';
-import VerticalLine from 'interface/others/charts/VerticalLine';
-import { formatDuration } from 'common/format';
-
-import './ManaLevelGraph.scss';
-
-class ManaChart extends React.PureComponent {
-  static propTypes = {
-    mana: PropTypes.arrayOf(PropTypes.shape({
-      x: PropTypes.number.isRequired,
-      y: PropTypes.number.isRequired,
-    })).isRequired,
-    deaths: PropTypes.arrayOf(PropTypes.shape({
-      x: PropTypes.number.isRequired,
-    })),
-    bossData: PropTypes.arrayOf(PropTypes.shape({
-      title: PropTypes.string.isRequired,
-      borderColor: PropTypes.string.isRequired,
-      backgroundColor: PropTypes.string.isRequired,
-      data: PropTypes.arrayOf(PropTypes.shape({
-        x: PropTypes.number.isRequired,
-        y: PropTypes.number.isRequired,
-      })).isRequired,
-    })).isRequired,
-    startTime: PropTypes.number.isRequired,
-    endTime: PropTypes.number.isRequired,
-  };
-
-  static defaultProps = {
-    deaths: [],
-  };
-
-  colors = {
-    mana: {
-      border: 'rgba(2, 109, 215, 0.6)',
-      background: 'rgba(2, 109, 215, 0.25)',
-    },
-    death: 'rgba(255, 0, 0, 0.8)',
-  };
-
-  render() {
-    const { mana, deaths, bossData, startTime, endTime } = this.props;
-
-    const xValues = [];
-    const yValues = [0, 25, 50, 75, 100];
-    for (let i = startTime; i < endTime; i += 30*1000) {
-      xValues.push(i);
-    }
-
-    return (
-      <XYPlot
-        height={400}
-        yDomain={[0, 100]}
-        margin={{
-          top: 30,
-        }}
-      >
-        <DiscreteColorLegend
-          orientation="horizontal"
-          items={[
-            ...bossData.map(boss => ({ title: boss.title, color: boss.borderColor })),
-            { title: 'Mana', color: this.colors.mana.border },
-            { title: 'Deaths', color: this.colors.death },
-          ]}
-        />
-        <XAxis tickValues={xValues} tickFormat={value => formatDuration((value - startTime) / 1000)} />
-        <YAxis tickValues={yValues} tickFormat={value => `${value}%`} />
-        <VerticalGridLines
-          tickValues={xValues}
-          style={{
-            strokeDasharray: 3,
-            stroke: 'white',
-            fill: 'white',
-          }}
-        />
-        <HorizontalGridLines
-          tickValues={yValues}
-          style={{
-            strokeDasharray: 3,
-            stroke: 'white',
-            fill: 'white',
-          }}
-        />
-        {bossData.map(boss => (
-          <AreaSeries
-            data={boss.data}
-            color={boss.backgroundColor}
-            stroke="transparent"
-          />
-        ))}
-        {bossData.map(boss => (
-          <LineSeries
-            data={boss.data}
-            color={boss.borderColor}
-            strokeWidth={2}
-          />
-        ))}
-        <AreaSeries
-          data={mana}
-          color={this.colors.mana.background}
-          stroke="transparent"
-        />
-        <LineSeries
-          data={mana}
-          color={this.colors.mana.border}
-          strokeWidth={2}
-        />
-        {deaths.map(({ x }) => (
-          <VerticalLine
-            value={x}
-            style={{
-              line: { background: this.colors.death },
-            }}
-          />
-        ))}
-      </XYPlot>
-    );
-  }
-}
+import ManaChart from './ManaChart';
 
 class Mana extends React.PureComponent {
   static propTypes = {
@@ -174,7 +47,15 @@ class Mana extends React.PureComponent {
       });
   }
 
-  get plot() {
+  render() {
+    if (!this.state.bossHealth) {
+      return (
+        <div>
+          Loading...
+        </div>
+      );
+    }
+
     const { start, end, manaUpdates } = this.props;
 
     const mana = manaUpdates.map(({ timestamp, current, max }) => {
@@ -203,31 +84,17 @@ class Mana extends React.PureComponent {
     });
 
     return (
-      <ManaChart
-        mana={mana}
-        deaths={deaths}
-        bossData={bossData}
-        startTime={start}
-        endTime={end}
-      />
-    );
-  }
-
-  render() {
-    if (!this.state.bossHealth) {
-      return (
-        <div>
-          Loading...
-        </div>
-      );
-    }
-
-    return (
       <div>
         Playing Arcane well typically involves managing your mana properly. Things such as not going OOM during Arcane Power, not letting your mana cap, and ensuring you end the fight with as little mana as possible will all help in improving your DPS.<br /><br />
 
         <div className="graph-container">
-          {this.plot}
+          <ManaChart
+            mana={mana}
+            deaths={deaths}
+            bossData={bossData}
+            startTime={start}
+            endTime={end}
+          />
         </div>
       </div>
     );
