+++ src/parser/shared/modules/features/RaidHealthTab/TabComponent/Graph.js
@@ -1,23 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {
-  FlexibleWidthXYPlot as XYPlot,
-  DiscreteColorLegend,
-  XAxis,
-  YAxis,
-  VerticalGridLines,
-  HorizontalGridLines,
-  AreaSeries,
-} from 'react-vis';
 
 import fetchWcl from 'common/fetchWclApi';
 
-import VerticalLine from 'interface/others/charts/VerticalLine';
-import { formatDuration } from 'common/format';
+import RaidHealthChart from './RaidHealthChart';
 
-import './RaidHealthChart.scss';
-
-const DEATH_COLOR = 'rgba(255, 0, 0, 0.8)';
 const CLASS_CHART_LINE_COLORS = {
   DeathKnight: 'rgba(196, 31, 59, 0.6)',
   Druid: 'rgba(255, 125, 10, 0.6)',
@@ -33,132 +20,6 @@ const CLASS_CHART_LINE_COLORS = {
   DemonHunter: 'rgba(163, 48, 201, 0.6)',
 };
 
-class RaidHealthChart extends React.Component {
-  static propTypes = {
-    players: PropTypes.arrayOf(PropTypes.shape({
-      title: PropTypes.string.isRequired,
-      borderColor: PropTypes.string.isRequired,
-      backgroundColor: PropTypes.string.isRequired,
-      data: PropTypes.arrayOf(PropTypes.shape({
-        x: PropTypes.number.isRequired,
-        y: PropTypes.number.isRequired,
-      })).isRequired,
-    })).isRequired,
-    deaths: PropTypes.arrayOf(PropTypes.shape({
-      x: PropTypes.number.isRequired,
-    })).isRequired,
-    startTime: PropTypes.number.isRequired,
-    endTime: PropTypes.number.isRequired,
-  };
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      players: this.props.players.map(player => ({ disabled: false, ...player })),
-    };
-    this.togglePlayer = this.togglePlayer.bind(this);
-  }
-
-  state = {
-    players: [],
-  };
-
-  togglePlayer(index) {
-    this.setState((prevState) => {
-      const players = prevState.players;
-      players[index].disabled = !players[index].disabled;
-      return {
-        players,
-      };
-    });
-  }
-
-  render() {
-    const { players, deaths, startTime, endTime } = this.props;
-
-    const xValues = [];
-    const yValues = [];
-    for (let i = 0; i < (endTime - startTime) / 1000; i += 30) {
-      xValues.push(i);
-    }
-    for (let i = 0; i <= players.length; i += 2) {
-      yValues.push(i * 100);
-    }
-
-    // XYPlot injects ton of properties into its children, I want a plain div
-    const LegendWrapper = (props) => (
-      <div className="legend-wrapper horizontal">
-        {props.children}
-      </div>
-    );
-
-    return (
-      <XYPlot
-        height={500}
-        yDomain={[0, players.length * 100]}
-        margin={{
-          left: 60,
-          top: 70,
-        }}
-        stackBy="y"
-      >
-        <LegendWrapper>
-          <DiscreteColorLegend
-            orientation="horizontal"
-            items={[
-              ...this.state.players.map(player => ({
-                title: player.title,
-                color: player.borderColor,
-                disabled: player.disabled,
-              })),
-              { title: 'Deaths', color: DEATH_COLOR },
-            ]}
-            onItemClick={(item, i) => {
-              if (item.title === 'Deaths') {
-                return;
-              }
-              this.togglePlayer(i);
-            }}
-          />
-        </LegendWrapper>
-        <XAxis tickValues={xValues} tickFormat={value => formatDuration(value)} />
-        <YAxis tickValues={yValues} tickFormat={value => `${value}%`} />
-        <VerticalGridLines
-          tickValues={xValues}
-          style={{
-            strokeDasharray: 3,
-            stroke: 'white',
-            fill: 'white',
-          }}
-        />
-        <HorizontalGridLines
-          tickValues={yValues}
-          style={{
-            strokeDasharray: 3,
-            stroke: 'white',
-            fill: 'white',
-          }}
-        />
-        {this.state.players.filter(player => !player.disabled).map(player => (
-          <AreaSeries
-            data={player.data}
-            color={player.backgroundColor}
-            stroke="transparent"
-            stack
-          />
-        ))}
-        {deaths.map(({ x }) => (
-          <VerticalLine
-            value={x}
-            style={{
-              line: { background: DEATH_COLOR },
-            }}
-          />
-        ))}
-      </XYPlot>
-    );
-  }
-}
 
 class Graph extends React.PureComponent {
   static propTypes = {
@@ -198,9 +59,17 @@ class Graph extends React.PureComponent {
       });
   }
 
-  get plot() {
-    const { start, end } = this.props;
+  render() {
     const data = this.state.data;
+    if (!data) {
+      return (
+        <div>
+          Loading...
+        </div>
+      );
+    }
+
+    const { start, end } = this.props;
 
     const players = data.series.filter(item => !!CLASS_CHART_LINE_COLORS[item.type]);
 
@@ -255,28 +124,13 @@ class Graph extends React.PureComponent {
     const deaths = Object.entries(deathsBySecond).filter(([_, value]) => !!value).map(([key]) => ({ x: Number(key) }));
 
     return (
-      <RaidHealthChart
-        players={playerHealth}
-        deaths={deaths}
-        startTime={start}
-        endTime={end}
-      />
-    );
-  }
-
-  render() {
-    const data = this.state.data;
-    if (!data) {
-      return (
-        <div>
-          Loading...
-        </div>
-      );
-    }
-
-    return (
       <div className="graph-container">
-        {this.plot}
+        <RaidHealthChart
+          players={playerHealth}
+          deaths={deaths}
+          startTime={start}
+          endTime={end}
+        />
       </div>
     );
   }
