+++ src/parser/shared/modules/features/RaidHealthTab/TabComponent/RaidHealthChart.scss
@@ -1,8 +1,8 @@
 .legend-wrapper.horizontal {
   position: absolute;
   top: -15px;
-  width: 75%;
-  left: 12.5%;
+  width: 90%;
+  left: 5%;
 
   .rv-discrete-color-legend.horizontal {
     white-space: nowrap;
